Reorganize code, add some comments. No code actual code changes
